Updated parser to latest commit. Fixed many small bugs in ICU. Added kwarg to ICU for sending user input file for analysis. Updated requirements.txt

+++ main.py
@@ -3,7 +3,8 @@ import requests
 import subprocess
 import shutil
 import os
-from pprint import pprint
+import argparse
+import traceback
 from rcon.source import Client
 from DataBase import DataBase
 from dotenv import load_dotenv
@@ -16,6 +17,7 @@ RCON_ADDRESS: str = os.getenv("RCON_ADDRESS")
 RCON_PORT: int = int(os.getenv("RCON_PORT"))
 RCON_PASSWD: str = os.getenv("RCON_PASSWD")
 GAME_SERVER_DIR: str = os.getenv("GAME_SERVER_DIR")
+PARSER_CONFIG: str = os.getenv("PARSER_CONFIG")
 db: DataBase = DataBase()
 
 
@@ -39,13 +41,27 @@ def getLatestDemoFile() -> str:
 def sendFile(fileName: str) -> None:
     filePath: str = os.path.join(os.path.dirname(__file__), "parser", "DemoFiles", "csv", fileName)
     with open(filePath, "rb") as fp:
-        req = requests.post(API_ENDPOINT, files={"csvFile": fp}, data={"apikey": API_KEY})
-        data = req.json()
-        if data["success"]:
-            print("Uploaded file for analysis successfully!")
-        else:
-            print("Failed uploading data")
-            print(f"Reason {data["message"]}")
+        try:
+            # req = requests.post("API_ENDPOINT", files={"csvFile": fp}, data={"apikey": API_KEY})
+            req = requests.post(f"{API_ENDPOINT}/{API_KEY}", files={"csvFile": fp})
+            if req.status_code == 403:
+                print("Invalid API KEY!")
+                return
+            if req.status_code != 200:
+                print(f"Something went wrong: [{req.status_code}]")
+                return
+            
+            data = req.json()
+            if data["success"]:
+                print("Uploaded file for analysis successfully!")
+            else:
+                print("Failed uploading data")
+                print(f"Reason {data['message']}")
+        except Exception:
+            print("Error sending...")
+            print(req.status_code)
+            print(req.text)
+            print(traceback.format_exc())
 
 
 def copyFileToParserBucket(fileName: str) -> None:
@@ -63,7 +79,7 @@ def deleteAllFilesInFolder(folderPath: str) -> None:
         try:
             os.remove(filePath)
             print(f"Deleted: {filePath}")
-        except Exception as e:
+        except Exception:
             print(f"Failed to delete {filePath}.\n Exception: {e}")
 
 
@@ -78,7 +94,7 @@ def parseFile() -> str:
     baseDir: str = os.path.dirname(__file__)
     parserDir: str = os.path.join(baseDir, "parser")
     os.chdir(parserDir)
-    args = ["py", "main.py"]
+    args = ["py", "main.py", "-cnf", f"./configs/{PARSER_CONFIG}"]
     subprocess.run(args)
     os.chdir(baseDir)
     allFiles: list[str] = os.listdir(os.path.join(parserDir, "DemoFiles", "csv"))
@@ -86,12 +102,37 @@ def parseFile() -> str:
     return targetFile
 
 
+def parseAndSend(fileName: str) -> None:
+    print(f"Target Acquired: {fileName}")
+    print("Copying file to parser bucket...")
+    copyFileToParserBucket(fileName)
+
+    try:
+        print("Parsing File...")
+        parsedFile: str = parseFile()
+    except Exception:
+        print("Parser Failed...")
+        print(traceback.format_exc())
+        return
+        
+    print(f"Sending file: {parsedFile}...")
+    try:
+        sendFile(parsedFile)
+    except Exception:
+        print(f"Failed to send file {parsedFile}")
+        print(traceback.format_exc())
+        return
+    
+    db.insert(fileName)
+    print("File Sent Successfully.")
+
+
 def runLoop():
     try:
         with Client(RCON_ADDRESS, RCON_PORT, passwd=RCON_PASSWD) as client:
             response: str = client.run('users')
-    except Exception as e:
-        print("Unable to connect to game server.")
+    except Exception:
+        print(f"Unable to connect to game server => {RCON_ADDRESS}:{RCON_PORT}")
         return
     
     userCount: int = int(response[-8:-6])
@@ -110,29 +151,36 @@ def runLoop():
         print("File already sent...")
         return
     
-    print(f"Target Acquired: {fileName}")
-    print("Copying file to parser bucket...")    
-    copyFileToParserBucket(fileName)
-
-    print("Parsing File...")
-    parsedFile: str = parseFile()
-
-    print(f"Sending file: {parsedFile}...")
     try:
-        sendFile(parsedFile)
-    except Exception as e:
-        print(f"Failed to send file {parsedFile}")
-        print(e)
-        return
-    
-    print("File Sent Successfully.")
-    print("Performing cleanup...")
-    deleteParserBucketFiles()
-    db.insert(fileName)
-    print("Cleanup Done.")
+        parseAndSend(fileName)
+    except Exception:
+        print(traceback.format_exc())
+    finally:
+        print("Performing cleanup...")
+        deleteParserBucketFiles()
+        print("Cleanup Done.")
 
 
 def main():
+    parser = argparse.ArgumentParser(description="Intermediatory Connection Utility for CS:GO and YAA-CS.")
+    parser.add_argument('-f', '--file', type=str, help="Absolute Demo File path.")
+
+    args = parser.parse_args()
+
+    filePath = args.file
+
+    if filePath and os.path.exists(filePath):
+        try:
+            parseAndSend(filePath)
+        except Exception:
+            print(traceback.format_exc())
+        finally:
+            print("Performing cleanup...")
+            deleteParserBucketFiles()
+            print("Cleanup Done.")
+        return
+    
+    print(f"Starting auto-detection mode.")
     while True:
         runLoop()
         time.sleep(10)
